requirements.txt
arxiv>=2.1.0
anthropic>=0.40.0
requests>=2.31.0
torch>=2.0.0
transformers>=4.40.0
numpy>=1.24.0
tqdm>=4.65.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
